requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
email-validator
python-dotenv
gunicorn
psycopg2